Split FormContext and actual Form into separate code files
