+++ app/webpacker/components/CompetitionForm/FormSections/NameDetails.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { InputString } from '../Inputs/FormInputs';
 import { useStore } from '../../../lib/providers/StoreProvider';
 import { competitionMaxShortNameLength } from '../../../lib/wca-data.js.erb';
-import { useFormObject } from '../../wca/FormProvider/EditForm';
+import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
 
 export default function NameDetails() {
   const { isPersisted, isAdminView } = useStore();
