+++ app/webpacker/components/CompetitionForm/FormSections/RegistrationDetails.js
@@ -9,7 +9,7 @@ import {
 import ConditionalSection from './ConditionalSection';
 import I18n from '../../../lib/i18n';
 import SubSection from '../../wca/FormProvider/SubSection';
-import { useFormObject } from '../../wca/FormProvider/EditForm';
+import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
 
 const guestsEnabledOptions = [true, false].map((bool) => ({
   value: bool,
