+++ app/webpacker/components/CompetitionForm/FormSections/RegistrationFees.js
@@ -10,7 +10,7 @@ import { calculateDuesUrl } from '../../../lib/requests/routes.js.erb';
 import useLoadedData from '../../../lib/hooks/useLoadedData';
 import ConditionalSection from './ConditionalSection';
 import SubSection from '../../wca/FormProvider/SubSection';
-import { useFormObject } from '../../wca/FormProvider/EditForm';
+import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
 
 const currenciesOptions = Object.keys(currenciesData.byIso).map((iso) => ({
   key: iso,
